Style upgrade 1

+++ src/app/components/searchresults.jsx
@@ -12,13 +12,15 @@ import logo from "@/app/images/logo.png";
 import Image from "next/image";
 
 export default function SearchResults(props){
-    const {cart} = useCart();
     const allProducts = props.products; //arreglo que contiene todos los productos, de acá se sacan los resultados
     const searchW = props.word; //palabra de busqueda
     const [newSearchW, setNewSearchW] = useState("default"); //nuevo valor de busqueda
-    const [displayCart, setDisplayCart] = useState(false); //estado del carrito, si esta abierto o cerrado
     const [results, setResults] = useState([...allProducts]); //resultados de la busqueda
     const [cantidadR, setCantidadR] = useState("default"); //cantidad de resultados
+
+    //Carrito
+    const {cart} = useCart();
+    const [displayCart, setDisplayCart] = useState(false); //estado del carrito, si esta abierto o cerrado
     
     //Filtros de busqueda
     const [orden, setOrden] = useState("Relevancia"); //orden de los productos
@@ -55,24 +57,43 @@ export default function SearchResults(props){
         // eslint-disable-next-line react-hooks/exhaustive-deps
         if(searchW === "empty"){
             setCantidadR(allProducts.length); //si la palabra de busqueda es vacia, se le asigna el valor de "empty" para que no haya errores
+            setNewSearchW(purged); //actualiza el valor de la nueva busqueda
+        }else if(searchW === "Maltas"){setResults(results.filter(x => x.type === "Malta")); setCantidadR(results.length); setNewSearchW("Maltas");
+        }else if(searchW === "Levaduras"){setResults(results.filter(x => x.type === "Levadura")); setCantidadR(results.length); setNewSearchW("Levaduras");
+        }else if(searchW === "Lupulos"){setResults(results.filter(x => x.type === "Lupulo")); setCantidadR(results.length); setNewSearchW("Lupulos");
+        }else if(searchW === "Kits"){setResults(results.filter(x => x.productType === "Kit")); setCantidadR(results.length); setNewSearchW("Kits");
+        }else if(searchW === "Equipamiento"){setResults(results.filter(x => x.productType === "Insumo")); setCantidadR(results.length); setNewSearchW("Equipamiento");
         }else{
             setCantidadR(results.length); //si la palabra de busqueda no es vacia, se le asigna el valor de la cantidad de resultados
+            setNewSearchW(purged); //actualiza el valor de la nueva busqueda
         } //si la palabra de busqueda es vacia, se le asigna el valor de "empty" para que no haya errores
-        setNewSearchW(purged); //actualiza el valor de la nueva busqueda
+        
     },[])
 
     useEffect(() => {
+    
     let filtered = [...allProducts];
 
     // Apply search term filter
     if (newSearchW !== "" && newSearchW !== "default" && newSearchW !== "empty") {
-        const purged = purgeSearch(newSearchW);
-        filtered = findResults(purged);
+        //Filtros de CirculoCategorias
+        if(newSearchW === "Maltas"){filtered = allProducts.filter(x => x.type === "Malta"); setIngredienteTipo("Malta");
+        }else if(newSearchW === "Levaduras"){filtered = allProducts.filter(x => x.type === "Levadura"); setIngredienteTipo("Levadura");
+        }else if(newSearchW === "Lupulos"){filtered = allProducts.filter(x => x.type === "Lupulo"); setIngredienteTipo("Lupulo");
+        }else if(newSearchW === "Kits"){filtered = allProducts.filter(x => x.productType === "Kit"); setProducto("Kit");
+        }else if(newSearchW === "Equipamiento"){filtered = allProducts.filter(x => x.productType === "Insumo"); setProducto("Insumo");
+        //================================================
+        }else{
+            const purged = purgeSearch(newSearchW);
+            filtered = findResults(purged); //solo purgamos acá, porque en la base de datos las categorías están en mayusculas.
+        }
     }
 
     // Apply other filters
     if (producto !== "default") filtered = filtered.filter(x => x.productType === producto);
     if (cervezaTipo !== "default") filtered = filtered.filter(x => x.productType === "Cerveza" && x.type === cervezaTipo);
+    if (ingredienteTipo !== "default") { filtered = filtered.filter(x => x.type === ingredienteTipo)};
+    if (tipoElemento !== "default") { filtered = filtered.filter(x => x.type === tipoElemento)};
     //if (cervezaCapacidad !== "default") filtered = filtered.filter(x => x.productType === "Cerveza" && x.quantity === cervezaCapacidad);
     //if (precioCerveza !== "default" ) {
     //    switch (precioCerveza) {
@@ -89,12 +110,6 @@ export default function SearchResults(props){
     //            break;
     //    }
     //}
-    if (ingredienteTipo !== "default") { filtered = filtered.filter(x => x.type === ingredienteTipo)};
-    if (tipoElemento !== "default") { filtered = filtered.filter(x => x.type === tipoElemento)};
-
-
-
-
 
     if(orden === "Relevancia"){
         filtered = filtered.sort((a, b) => b.priority - a.priority); // Ordena por prioridad
@@ -113,8 +128,8 @@ export default function SearchResults(props){
 
     return(
         <div>
-            <div className="w-full theme2 fixed top-0 z-25">
-                <div className="grid grid-flow-col min-[1000px]:grid-cols-3 place-self-start min-[1000px]:place-content-center  text-center place-items-start min-[1000px]:place-items-center h-[100px] px-2 font-bold text-4xl w-[500px] min-[650px]:w-[650px] min-[800px]:w-[800px] min-[1000px]:w-[1100px] min-[1400px]:w-[1400px] mx-auto gap-4 pt-4 min-[1000px]:pt-1">
+            <div className="w-full theme2 fixed top-0 z-25 min-w-[600px]">
+                <div className="grid grid-flow-col min-[1000px]:grid-cols-3 place-self-start min-[1000px]:place-content-center text-center place-items-start min-[1000px]:place-items-center h-[80px] px-2 font-bold text-4xl w-[500px] min-[650px]:w-[650px] min-[800px]:w-[800px] min-[1000px]:w-[1100px] min-[1400px]:w-[1400px] mx-auto gap-4 pt-4 min-[1000px]:pt-1">
                     {/* Otro Navbar, este necesita un buscador diferente <Link href="/talleres/"><h1 className="TextShine Bigger">Talleres</h1></Link> para que no se recarge la pagina al buscar en la misma */}
                     <div>
                         <Image
@@ -122,33 +137,34 @@ export default function SearchResults(props){
                             alt="no image"
                             width={400}
                             height={400}
-                            className="mx-auto w-[65px] h-[65px]"
+                            className="mx-auto w-[60px] h-[50px]"
                         />
                     </div>
                     <div className="grid grid-flow-row place-content-center place items-center gap-2">
                         <div className="grid grid-flow-col place-content-center place-self-center place-items-center align-middle">
                             <SearchAgain set={setNewSearchW}/>
                         </div>
-                        <div className="grid grid-flow-col place-self-center place-content-center place-items-center text-center gap-4 min-[850px]:gap-8 text-sm min-[650px]:text-lg min-[850px]:text-xl">
-                            <Link href="/"><h1 className="TextShine Bigger">Inicio</h1></Link>
-                            <Link href="/search/empty"><h1 className="TextShine Bigger">Productos</h1></Link>
-                            <Link href="/about"><h1 className="TextShine Bigger">Sobre Nosotros</h1></Link>
-                        </div>
                     </div>
                     <div>
                         <CartButton setDisplay={setDisplayCart} display={displayCart} cart={cart}/>
                     </div>           
                 </div>
+                <div className="grid grid-flow-col theme4 w-full gap-8 place-self-center place-content-center place-items-center text-center text-xl p-2">
+                    <Link href="/"><h1 className="TextShine Bigger">Inicio</h1></Link>
+                    <Link href="/search/empty"><h1 className="TextShine Bigger">Productos</h1></Link>
+                    <Link href="/about"><h1 className="TextShine Bigger">Sobre Nosotros</h1></Link>
+                </div>
+
             </div>
 
-            <div className="grid grid-flow-col pt-[100px]">
+            <div className="grid grid-flow-col place-self-start min-[650px]:place-self-center place-content-center place-items-center pt-[130px]">
                 {/* Barra de filtros / listado de productos / carrito (cuando se activa) */}
-                <div className="grid grid-flow-col p-2">
-                    <div className="grid grid-flow-row place-self-start text-left pl-2 min-[1100px]:pl-4 min-[1700px]:pl-8 pt-2 gap-4">
+                <div className="grid grid-flow-col place-self-center place-content-start p-2">
+                    <div className="grid grid-flow-row place-self-start text-left pt-2 gap-4 w-[180px] min-[750px]:w-[275px] min-[1200px]:w-[300px] min-[1400px]:w-[350px]">
                         <div className="text-lg min-[1200px]:text-xl min-[1400px]:text-2xl min-[1700px]:text-3xl font-extrabold">Filtros de Busqueda</div>
                         <div>
                             <ButtonF head={true} set={setProducto} global={producto} value={"Cerveza"} text={"Cervezas"} setSearch={setNewSearchW}/>
-                            <div className="grid grid-flow-row place-content-start place-items-start pl-2 min-[1100px]:pl-4 min-[1400px]:pl-8">
+                            <div className="grid grid-flow-row place-content-start place-items-start pl-5">
                                 <ButtonF set={setCervezaTipo} global={cervezaTipo} value={"Amber"} text={"Amber"} setSearch={setNewSearchW}/>
                                 <ButtonF set={setCervezaTipo} global={cervezaTipo} value={"APA"} text={"APA"} setSearch={setNewSearchW}/>
                                 <ButtonF set={setCervezaTipo} global={cervezaTipo} value={"Blonde"} text={"Blonde"} setSearch={setNewSearchW}/>
@@ -162,7 +178,7 @@ export default function SearchResults(props){
                         </div>
                         <div>
                             <ButtonF head={true} set={setProducto} global={producto} value={"Ingrediente"} text={"Materias Primas"} setSearch={setNewSearchW}/>
-                            <div className="grid grid-flow-row place-content-start place-items-start pl-2 min-[1100px]:pl-4 min-[1400px]:pl-8">
+                            <div className="grid grid-flow-row place-content-start place-items-start pl-5">
                                 <ButtonF set={setIngredienteTipo} global={ingredienteTipo} value={"Lupulo"} text={"Lúpulo"} setSearch={setNewSearchW}/>
                                 <ButtonF set={setIngredienteTipo} global={ingredienteTipo} value={"Malta"} text={"Maltas"} setSearch={setNewSearchW}/>
                                 <ButtonF set={setIngredienteTipo} global={ingredienteTipo} value={"Levadura"} text={"Levaduras"} setSearch={setNewSearchW}/>
@@ -171,7 +187,7 @@ export default function SearchResults(props){
                         </div>
                         <div>
                             <ButtonF head={true} set={setProducto} global={producto} value={"Insumo"} text={"Equipamiento"} setSearch={setNewSearchW} />
-                            <div className="grid grid-flow-row place-content-start place-items-start pl-2 min-[1100px]:pl-4 min-[1400px]:pl-8">
+                            <div className="grid grid-flow-row place-content-start place-items-start pl-5">
                                 <ButtonF set={setTipoElemento} global={tipoElemento} value={"Instrumento"} text={"Medición"} setSearch={setNewSearchW}/>
                                 <ButtonF set={setTipoElemento} global={tipoElemento} value={"Envase"} text={"Envases y Tapas"} setSearch={setNewSearchW}/>
                                 <ButtonF set={setTipoElemento} global={tipoElemento} value={"Limpieza"} text={"Limpieza"} setSearch={setNewSearchW}/>
@@ -184,10 +200,10 @@ export default function SearchResults(props){
                     </div>
                     
 
-                    <div>
-                        <div className="grid grid-flow-col pb-4">
+                    <div className="grid w-[320px] min-[650px]:w-[440px] min-[1100px]:w-[800px] min-[1500px]:w-[1100px] min-[1700px]:w-[1200px]">
+                        <div className="grid grid-flow-row min-[650px]:grid-flow-col pb-4">
                             <div className="font-bold text-sm min-[650px]:text-base min-[900px]:text-lg pl-2">{cantidadR} productos encontrados</div>
-                            <select className="grid place-self-end theme6 p-2 w-[50%]" value={orden} onChange={(e) => setOrden(e.target.value)}>
+                            <select className="grid place-self-start min-[650px]:place-self-end theme6 p-2 w-[80%] min-[1050px]:w-[50%]" value={orden} onChange={(e) => setOrden(e.target.value)}>
                                 <option value="Relevancia">Relevancia</option>
                                 <option value="Mayor">Mayor a menor precio</option>
                                 <option value="Menor">Menor a mayor precio</option>
